+++ src/layman/user/util.py
@@ -1,8 +1,12 @@
 from flask import g
 from layman import LaymanError, authn
 from layman.authn import get_open_id_claims, get_iss_id, get_sub, is_user_with_name
-from layman.util import slugify, to_safe_names, check_workspace_name, get_workspaces, ensure_whole_user, delete_whole_user
+from layman.util import slugify, to_safe_names, check_workspace_name, get_workspaces, ensure_whole_user, delete_whole_user, delete_publications, get_publication_infos, patch_publication
 from layman.authn import redis as authn_redis, prime_db_schema as authn_prime_db_schema
+from layman.layer import LAYER_TYPE, util as layer_util
+from layman.map import MAP_TYPE, util as map_util
+from layman.authz import is_user
+from layman_settings import RIGHTS_EVERYONE_ROLE
 
 
 def get_user_profile(user_obj):
@@ -100,3 +104,131 @@ def get_username_suggestions_from_claims(claims):
 def delete_user(username=None):
     authn_redis.delete_user(username)
     delete_whole_user(username)
+
+
+def check_if_is_only_rw(access_rights, username):
+    read_set = set(access_rights['read'])
+    write_set = set(access_rights['write'])
+    if RIGHTS_EVERYONE_ROLE in read_set or RIGHTS_EVERYONE_ROLE in write_set:
+        return False
+    read_users = {u for u in read_set if is_user(u)}
+    write_users = {u for u in write_set if is_user(u)}
+    return (read_users == {username}) and (write_users == {username})
+
+
+def delete_user_public_publications(username):
+    def has_only_roles_without_everyone(user_list):
+        return all(not is_user(u) and u != RIGHTS_EVERYONE_ROLE for u in user_list)
+
+    skipped_publications = []
+    layers_to_be_deleted = []
+    maps_to_be_deleted = []
+    publications_to_be_patched = []
+    result = get_publication_infos(
+        context={
+            'actor_name': username,
+            'access_type': 'read',
+        },
+    )
+
+    for (workspace, publication_type, publication_name), pubinfo in result.items():
+        is_public = pubinfo['_is_public_workspace']
+        access_rights = pubinfo['access_rights']
+        geodata_type = pubinfo['geodata_type']
+        is_only_rw = check_if_is_only_rw(access_rights, username)
+        publication_name = pubinfo['name']
+        uuid = pubinfo["uuid"]
+
+        if is_public and is_only_rw:
+            if publication_type == LAYER_TYPE:
+                layers_to_be_deleted.append(((workspace, publication_type, publication_name), pubinfo))
+            elif publication_type == MAP_TYPE:
+                maps_to_be_deleted.append(((workspace, publication_type, publication_name), pubinfo))
+            else:
+                raise NotImplementedError(f"Unknown publication_type: {publication_type}")
+
+        elif is_public and not is_only_rw:
+            new_read = [u for u in access_rights['read'] if u != username]
+            new_write = [u for u in access_rights['write'] if u != username]
+            if (
+                    username in access_rights['read']
+                    and has_only_roles_without_everyone(new_write)
+            ):
+                skipped_publications.append({
+                    "workspace": workspace,
+                    "type": publication_type,
+                    "name": publication_name,
+                    "uuid": uuid,
+                })
+                continue
+
+            publications_to_be_patched.append(
+                (workspace, publication_name, publication_type, pubinfo, new_read, new_write, geodata_type)
+            )
+    if skipped_publications:
+        raise LaymanError(58, {
+            'message': f"User {username} cannot be removed from some publications because other users have read rights.",
+            'username': username,
+            'unable_to_delete_publications': skipped_publications,
+        })
+
+    for (workspace, publication_name, publication_type, pubinfo, new_read, new_write, geodata_type) in publications_to_be_patched:
+        kwargs = {
+            "access_rights": {
+                "read": new_read,
+                "write": new_write
+            },
+            "actor_name": username,
+            "external_table_uri": pubinfo.get('external_table_uri'),
+            "geodata_type": geodata_type,
+            "title": pubinfo.get('title'),
+            "description": pubinfo.get('description'),
+            "original_data_source": pubinfo.get('original_data_source'),
+            "is_part_of_user_delete": True,
+        }
+
+        patch_publication(
+            workspace,
+            publication_name,
+            publication_type,
+            layer_util.patch_layer if publication_type == LAYER_TYPE else map_util.patch_map,
+            layer_util.is_layer_chain_ready if publication_type == LAYER_TYPE else map_util.is_map_chain_ready,
+            kwargs,
+            {
+                "stop_sync_at": None,
+                "start_async_at": None,
+                "only_sync": True
+            } if publication_type == LAYER_TYPE else {
+                "start_at": None,
+                "only_sync": True
+            }
+        )
+
+    if layers_to_be_deleted:
+        delete_publications(
+            None,
+            None,
+            layer_util.is_layer_chain_ready,
+            layer_util.abort_layer_chain,
+            layer_util.delete_layer,
+            'delete',
+            'rest_workspace_layer.get',
+            'layername',
+            actor_name=username,
+            publications=layers_to_be_deleted
+        )
+    if maps_to_be_deleted:
+        delete_publications(
+            None,
+            None,
+            map_util.is_map_chain_ready,
+            map_util.abort_map_chain,
+            map_util.delete_map,
+            'delete',
+            'rest_workspace_map.get',
+            'mapname',
+            actor_name=username,
+            publications=maps_to_be_deleted
+        )
+
+    return skipped_publications